requirements.txt
python3
python-ply==3.10
pydot==1.2.4

